Refactor manager_requests into IPC module
