+++ astoria/consumers/astctl/metadata/set.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import click
 
-from astoria.common.manager_requests import MetadataSetManagerRequest
+from astoria.common.ipc import MetadataSetManagerRequest
 from astoria.consumers.astctl.command import Command
 
 loop = asyncio.get_event_loop()
