+++ astoria/consumers/astctl/static_disks/remove_all.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import click
 
-from astoria.common.manager_requests import RemoveAllStaticDisksRequest
+from astoria.common.ipc import RemoveAllStaticDisksRequest
 from astoria.consumers.astctl.command import Command
 
 loop = asyncio.get_event_loop()
