+++ astoria/managers/astmetad/metadata_manager.py
@@ -4,11 +4,8 @@ import asyncio
 import logging
 from typing import Dict, List, Optional, Set, Tuple, Type
 
+from astoria.common.ipc import MetadataSetManagerRequest, RequestResponse
 from astoria.common.manager import StateManager
-from astoria.common.manager_requests import (
-    MetadataSetManagerRequest,
-    RequestResponse,
-)
 from astoria.common.messages.astdiskd import DiskInfo, DiskType, DiskUUID
 from astoria.common.messages.astmetad import Metadata, MetadataManagerMessage
 from astoria.managers.mixins.disk_handler import DiskHandlerMixin
